refactor(apps): update component hierarchy - update interfaces
