+++ apps/moderation/src/components/error-cards/index.tsx
@@ -1,0 +1,5 @@
+import NoItemsFound from './no-items-found';
+import PromptAuthorization from './prompt-authorization';
+import PromptAuthentication from './prompt-authentication';
+
+export { NoItemsFound, PromptAuthentication, PromptAuthorization };
