+++ apps/moderation/src/components/error-cards/prompt-authorization.tsx
